+++ src/FwdState.h
@@ -10,7 +10,7 @@
 #define SQUID_FORWARD_H
 
 #include "base/forward.h"
-#include "base/JobWait.h"
+#include "async/JobWait.h"
 #include "base/RefCount.h"
 #include "clients/forward.h"
 #include "comm.h"
