+++ src/comm/Connection.cc
@@ -7,7 +7,7 @@
  */
 
 #include "squid.h"
-#include "base/JobWait.h"
+#include "async/JobWait.h"
 #include "CachePeer.h"
 #include "cbdata.h"
 #include "comm.h"
