+++ src/log/TcpLogger.h
@@ -10,7 +10,7 @@
 #define _SQUID_SRC_LOG_TCPLOGGER_H
 
 #include "async/AsyncJob.h"
-#include "base/JobWait.h"
+#include "async/JobWait.h"
 #include "comm/forward.h"
 #include "ip/Address.h"
 #include "log/forward.h"
